Rename ThreadModeTestBase back
