+++ resilience4j-core/src/test/java/io/github/resilience4j/core/ThreadModeExtensionTest.java
@@ -10,7 +10,7 @@ import static org.assertj.core.api.Assertions.assertThat;
  * {@code @TestTemplate} method for both platform and virtual thread modes.
  */
 @ExtendWith(ThreadModeExtension.class)
-class ThreadModeTestBaseParameterizedIntegrationTest {
+class ThreadModeExtensionTest {
 
     @TestTemplate
     void shouldRunWithBothThreadModes(ThreadType threadType) {
